Add animation and render pass pipeline,Add a simple skeleton and animation system,Organize relevant code

+++ Witchcraft/WitchcraftEngine/SRC/String/SStringUtils.h
@@ -35,7 +35,7 @@ namespace SString
 	static std::wstring StringToWstring(const std::string& str)
 	{
 		std::wstring result;
-		size_t len = MultiByteToWideChar(CP_ACP, 0, str.c_str(), str.size(), NULL, 0);
+		size_t len = MultiByteToWideChar(CP_ACP, 0, str.c_str(), (int)str.size(), NULL, 0);
 		if (len < 0)return result;
 		wchar_t* buffer = new wchar_t[len + 1];
 		if (buffer == NULL)return result;
@@ -50,7 +50,7 @@ namespace SString
 	static std::string WstringToString(const std::wstring& wstr)
 	{
 		std::string result;
-		size_t len = WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), wstr.size(), NULL, 0, NULL, NULL);
+		size_t len = WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), (int)wstr.size(), NULL, 0, NULL, NULL);
 		if (len <= 0)return result;
 		char* buffer = new char[len + 1];
 		if (buffer == NULL)return result;
